integrate unique code generation with admin portal

+++ src/main/java/com/bitspondon/quiz/domain/entities/Chapter.java
@@ -34,6 +34,6 @@ public class Chapter {
     private Subject subject;
 
     @Transient
-    private int subjectId;
+    private Long subjectId;
 }
